requirements.txt
Flask
gTTS
google-generativeai
googletrans
gunicorn
